+++ include/summit/app/grid/pipeline/chip.hpp
@@ -119,6 +119,7 @@ struct Chip {
                                         fov_id, task.um2px_r()
                                     )
                                   );
+                summit::grid::log.debug("white channel, fov id:({}, {}) theta: {}", fov_id.x, fov_id.y, theta);
                 // std::cout << "white channel detect theta " 
                 //     << fov_id << ": " << theta << std::endl;
                 cv::Mat mat_loc = mat.clone();
@@ -138,6 +139,7 @@ struct Chip {
                     task.mk_hd_um(),
                     nucleona::stream::null_out
                 );
+                summit::grid::log.debug("white channel, fov id:({}, {}) um2px_r: {}", fov_id.x, fov_id.y, um2px_r);
                 mk_regs = aruco_mk_det(mat_loc, fov_id, um2px_r);
                 mk_regs = __alias::cmk_det::reg_mat_infer(mk_regs, mk_num.y, mk_num.x);
 
@@ -265,6 +267,94 @@ struct Chip {
         task.set_um2px_r(best_um2px_r);
         return true;
     }
+    template<class T>
+    void draw_grid(
+        cv::Mat& mat, 
+        const std::vector<T>& gl_x, 
+        const std::vector<T>& gl_y, 
+        int line_value
+    ) const {
+        for(auto&& xl : gl_x) {
+            cv::line(mat,
+                cv::Point(xl, 0), 
+                cv::Point(xl, mat.rows),
+                line_value
+            );
+        }
+        for(auto&& yl : gl_y) {
+            cv::line(mat,
+                cv::Point(0, yl), 
+                cv::Point(mat.cols, yl),
+                line_value
+            );
+        }
+    }
+    template<class T>
+    cv::Mat view_grid_raw_img(const chipimgproc::GridRawImg<T>& grm, int line_color) const {
+        cv::Mat res = grm.mat().clone();
+        for(auto&& xl : grm.gl_x()) {
+            cv::line(res,
+                cv::Point(xl, 0), 
+                cv::Point(xl, grm.rows()),
+                cv::Scalar(line_color)
+            );
+        }
+        for(auto&& yl : grm.gl_y()) {
+            cv::line(res,
+                cv::Point(0, yl), 
+                cv::Point(grm.cols(), yl),
+                cv::Scalar(line_color)
+            );
+        }
+        return res;
+    }
+    void gridline_debug_image_proc(model::Task& task) const { 
+        /**
+         * generate additional debug grid line image
+         * 1. white channel stitch
+         *      1. rotate
+         * 2. each stitched image grid line draw
+         */
+        chipimgproc::stitch::GridlineBased gl_stitcher;
+        std::map<std::string, const nlohmann::json*> channel_params;
+        std::string wh_name;
+        for(auto&& ch : task.channels()) {
+            channel_params[ch.at("name").get<std::string>()] = &ch;
+            if(ch.at("filter").get<int>() == 0) {
+                wh_name = ch.at("name").get<std::string>();
+            }
+        }
+        std::vector<int> filter_to_color({0, 0, 1, 0, 2, 0});
+        auto tpl_mtm = task.multi_tiled_mat().begin()->second.value();
+        for(auto&& [fov_id, img_data] : task.white_channel_imgs()) {
+            auto& [path, img] = img_data;
+            cv::Mat img_loc = img.clone();
+            rotate_calibrator(img_loc, task.rot_degree().value());
+            auto& tpl_gri = tpl_mtm.get_fov_img(fov_id.x, fov_id.y);
+            tpl_gri.mat() = img_loc;
+        }
+        auto gl_wh_stitch = gl_stitcher(tpl_mtm);
+        task.set_stitched_img(wh_name, std::move(gl_wh_stitch));
+
+        std::vector<cv::Mat> channels;
+        channels.resize(task.stitched_img().size());
+        for(auto& [chid, st_img] : task.stitched_img()) {
+            auto loc_st_img = st_img.mat().clone();
+            if(chid == wh_name) {
+                cv::Mat tmp;
+                loc_st_img.convertTo(tmp, CV_16UC1, 256);
+                loc_st_img = tmp;
+            } else {
+                loc_st_img = chipimgproc::viewable(loc_st_img, 5000);
+            }
+            draw_grid(loc_st_img, st_img.gl_x(), st_img.gl_y(), 32767);
+            auto ch_filter = channel_params.at(chid)->at("filter").get<int>();
+            channels.at(filter_to_color.at(ch_filter)) = loc_st_img;
+        }
+        cv::Mat stitched_grid_all;
+        cv::merge(channels, stitched_grid_all);
+        cv::imwrite(task.debug_stitch().string(), stitched_grid_all);
+    }
     decltype(auto) operator()(model::Task& task) const {
         using namespace __alias;
         auto& model = task.model();
@@ -274,6 +364,7 @@ struct Chip {
                 auto du_ms = std::chrono::duration_cast<std::chrono::milliseconds>(du).count();
                 task.set_proc_time(du_ms / 1000.0);
             });
+            task.grid_log()["date"] = summit::utils::datetime("%Y/%m/%d %H:%M:%S", std::chrono::system_clock::now());
             auto& wh_ch_log = task.grid_log()["white_channel_proc"];
             if(task.model().auto_gridding()) {
                 if(!white_channel_proc(task)) {
@@ -326,6 +417,10 @@ struct Chip {
             task.grid_log()["version"] = summit::grid::version().to_string();
             task.summary_channel_log();
             task.model().heatmap_writer().flush();
+            if(task.model().debug() >= 5) {
+                gridline_debug_image_proc(task);
+            }
+
         } catch( const std::exception& e ) {
             task.set_grid_done(false);
             task.grid_log()["grid_fail_reason"] = e.what();